+++ app/entities/service_name/admins_service_name.py
@@ -1,6 +1,6 @@
 from sqladmin import ModelView
 
-from entities.service_name.model_service_name import ServiceNameModel
+from entities.service_name.models_service_name import ServiceNameModel
 
 
 class ServiceNameAdmin(ModelView, model=ServiceNameModel):
