+++ app/entities/service_name/schemas_service_name.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from pydantic import constr, condecimal
 
-from entities.service_name.model_service_name import ServiceNameInSchema
+from entities.service_name.models_service_name import ServiceNameInSchema
 
 
 class ServiceNameOptionalSchema(ServiceNameInSchema):
