+++ app/models/offer_model.py
@@ -8,7 +8,7 @@ from sqlalchemy import UniqueConstraint
 from models.base_models import BaseSQLModel, DateCreatedChangedBase
 
 if TYPE_CHECKING:
-    from entities.service_name.model_service_name import ServiceNameModel
+    from entities.service_name.models_service_name import ServiceNameModel
     from entities.users.models_user import EmployeeModel
