+++ app/routers/order_routers.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm import joinedload, selectinload
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from models import OrderModel
-from entities.service_name.model_service_name import ServiceNameModel
+from entities.service_name.models_service_name import ServiceNameModel
 from models.choices import StatusOrder
 from routers.consts import RouteSlug
 from models.database import get_session
