+++ app/routers/service_routers.py
@@ -1,11 +1,11 @@
 from fastapi import Depends, APIRouter, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from entities.service_name.model_service_name import ServiceNameModel
+from entities.service_name.models_service_name import ServiceNameModel
 from routers.consts import RouteSlug
 from models.database import get_session
 from services.service_service import ServiceNameService
-from entities.service_name.schema_service_name import ServiceNameOptionalSchema
+from entities.service_name.schemas_service_name import ServiceNameOptionalSchema
 from dependencies.base_dependency import (
     ValidGetByIdDependency,
     valid_empty_schema,
