+++ app/services/service_service.py
@@ -1,7 +1,7 @@
 from typing import Type
 
 from models import OfferLinkModel
-from entities.service_name.model_service_name import ServiceNameModel
+from entities.service_name.models_service_name import ServiceNameModel
 from services.base_service import AbstractService
